refactor: generate single typescript script for vue file
