--- packages/vue-code-gen/package.json
@@ -1,25 +1,0 @@
-{
-	"name": "@volar/vue-code-gen",
-	"version": "0.39.5",
-	"main": "out/index.js",
-	"license": "MIT",
-	"files": [
-		"out/**/*.js",
-		"out/**/*.d.ts"
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/johnsoncodehk/volar.git",
-		"directory": "packages/vue-code-gen"
-	},
-	"devDependencies": {
-		"typescript": "latest"
-	},
-	"dependencies": {
-		"@volar/code-gen": "0.39.5",
-		"@volar/source-map": "0.39.5",
-		"@vue/compiler-core": "^3.2.37",
-		"@vue/compiler-dom": "^3.2.37",
-		"@vue/shared": "^3.2.37"
-	}
-}
