--- packages/vue-code-gen/tsconfig.build.json
@@ -1,25 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "out",
-		"rootDir": "src",
-	},
-	"include": [
-		"src"
-	],
-	"exclude": [
-		"node_modules",
-		".vscode-test"
-	],
-	"references": [
-		{
-			"path": "../shared/tsconfig.build.json"
-		},
-		{
-			"path": "../source-map/tsconfig.build.json"
-		},
-		{
-			"path": "../code-gen/tsconfig.build.json"
-		},
-	],
-}
