+++ packages/vue-language-core/src/index.ts
@@ -4,5 +4,10 @@ export * from './sourceFile';
 export * from './documentRegistry';
 export * from './types';
 export * from './lsContext';
+export * as refSugarRanges from './parsers/refSugarRanges';
+export * as scriptRanges from './parsers/scriptRanges';
+export * as scriptSetupConvertRanges from './parsers/scriptSetupConvertRanges';
+export * as scriptSetupRanges from './parsers/scriptSetupRanges';
 export * as localTypes from './utils/localTypes';
 export * as tsShared from './utils/ts';
+export * from './generators/template';
