+++ packages/vue-language-core/src/parsers/refSugarRanges.ts
