+++ packages/vue-language-core/src/parsers/scriptRanges.ts
