+++ packages/vue-language-core/src/parsers/scriptSetupConvertRanges.ts
