+++ packages/vue-language-core/src/parsers/scriptSetupRanges.ts
