+++ packages/vue-language-core/src/utils/vue2TemplateCompiler.ts
