+++ packages/vue-language-service/tsconfig.build.json
@@ -28,9 +28,6 @@
 			"path": "../typescript-language-service/tsconfig.build.json"
 		},
 		{
-			"path": "../vue-code-gen/tsconfig.build.json"
-		},
-		{
 			"path": "../vue-language-core/tsconfig.build.json"
 		},
 		{
