+++ packages/vue-test-workspace/rename/component-tag/input/component.vue
@@ -1,0 +1,2 @@
+<script lang="ts" setup>
+</script>
