+++ packages/vue-test-workspace/rename/component-tag/output/component.vue
@@ -1,0 +1,2 @@
+<script lang="ts" setup>
+</script>
