+++ packages/vue-test-workspace/rename/regular-component/input/entry.vue
@@ -1,0 +1,18 @@
+import { defineComponent } from 'vue';
+<template>
+   {{ foo }}
+ <!-- ^^^rename: bar -->
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+   setup() {
+      return {
+         foo: 1,
+      // ^^^rename: bar
+      };
+   },
+});
+</script>
