+++ packages/vue-test-workspace/rename/regular-component/output/entry.vue
@@ -1,0 +1,18 @@
+import { defineComponent } from 'vue';
+<template>
+   {{ bar }}
+ <!-- ^^^rename: bar -->
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+   setup() {
+      return {
+         bar: 1,
+      // ^^^rename: bar
+      };
+   },
+});
+</script>
