+++ packages/vue-test-workspace/vue-tsc/reference-type-in-template/main.vue
@@ -1,0 +1,10 @@
+<template>
+{{
+    // @ts-expect-error
+    1 as Foo
+}}
+</template>
+
+<script lang="ts" setup>
+type Foo = string;
+</script>
